Create pinging.ps1
Create a script that simulates pinging a remote server. It takes two named parameters. The first parameter is IP-address of a server, second how many times the ping is tried. The script also ask the permission to make ping with. Pings are made only if a user says Yes. Give default values for parameters, so if the script will be run without arguments, it will try to ping 192.168.1.1 only once.